Fix for warning on Linux PPC/gcc 4.1.2
git-svn-id: svn://scm.orgis.org/mpg123/trunk@1122 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/libmpg123.c
@@ -575,7 +575,7 @@ int mpg123_decode(mpg123_handle *mh,unsigned char *inmemory, size_t inmemsize, u
 		{
 			/* get what is needed - or just what is there */
 			int a = mh->buffer.fill > (outmemsize - mdone) ? outmemsize - mdone : mh->buffer.fill;
-			debug4("buffer fill: %i; copying %i (%i - %li)", (int)mh->buffer.fill, a, (int)outmemsize, *done);
+			debug4("buffer fill: %i; copying %i (%i - %li)", (int)mh->buffer.fill, a, (int)outmemsize, (long)*done);
 			memcpy(outmemory, mh->buffer.p, a);
 			/* less data in frame buffer, less needed, output pointer increase, more data given... */
 			mh->buffer.fill -= a;
